edit-content: split functions
